requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
torch==2.1.0
torchvision==0.16.0
pillow==10.1.0
python-multipart
# Installation directe depuis GitHub pour CLIP
git+https://github.com/openai/CLIP.git